requirements.txt
fastapi[all]
alembic
sqladmin
sqlalchemy
python-decouple
pre-commit
psycopg2-binary
passlib
bcrypt
redis
celery
flower
